att-script

+++ script-automas-criarsys.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 SERVICE_FILE1="/etc/systemd/system/nginx-log-monitor.service"
+PATH_ENV="/etc/nginx-monitor-env"
 
 if [ -f "$SERVICE_FILE1" ] ; then
   echo "INFO:  o arquivo de configuração já foi encontrado."
@@ -10,6 +11,18 @@ if [ -f "$SERVICE_FILE1" ] ; then
   exit 0
 fi
 
+if [ -f "$PATH_ENV" ] || [ -z "$DISCORD_WEBHOOK_URL" ]; then
+  echo "INFO:  o arquivo de ambiente foi encontrado ou a variável DISCORD_WEBHOOK_URL não está definida."
+  echo "      - Script: $PATH_ENV"
+  echo "INFO: Criando o arquivo de ambiente."
+  exit 0
+fi
+
+cat > "$PATH_ENV" << EOF
+DISCORD_WEBHOOK_URL=$DISCORD_WEBHOOK_URL
+EOF
+
+
 cat > "$SERVICE_FILE1" << EOF
 [Unit]
 Description=Real-time Nginx Error Log Monitor
@@ -17,7 +30,7 @@ After=network.target
 
 [Service]
 # O caminho para o nosso novo script de monitoramento
-ExecStart=/usr/local/bin/monitor_nginx_log.sh
+ExecStart=/home/ubuntu/automacao-conf-nginx-discord/monitor_nginx_log.sh
 
 # Política de reinicialização: se o script falhar por qualquer motivo, reinicie-o.
 Restart=always
@@ -26,7 +39,7 @@ Restart=always
 User=ubuntu
 
 # Arquivo que contém a variável de ambiente com o segredo
-EnvironmentFile=/etc/nginx-monitor-env
+EnvironmentFile=/home/ubuntu/automacao-conf-nginx-discord/monitor_nginx_log_env
 
 [Install]
 WantedBy=multi-user.target
